Support JDK17 for CI enviroment #13868
