+++ dubbo-api-docs/dubbo-api-docs-core/pom.xml
@@ -38,6 +38,7 @@
         <dependency>
             <groupId>org.apache.dubbo</groupId>
             <artifactId>dubbo</artifactId>
+            <version>3.2.0</version>
             <scope>compile</scope>
         </dependency>
