+++ dubbo-api-docs/dubbo-api-docs-examples/examples-provider-sca/pom.xml
@@ -39,7 +39,7 @@
         <dependency>
             <groupId>org.apache.dubbo.extensions.examples.apidocs</groupId>
             <artifactId>examples-api</artifactId>
-            <version>${parent.version}</version>
+            <version>${project.parent.version}</version>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
