+++ dubbo-api-docs/pom.xml
@@ -33,7 +33,7 @@
     <inceptionYear>2020</inceptionYear>
 
     <properties>
-        <project.build.jdkVersion>1.8</project.build.jdkVersion>
+        <project.build.jdkVersion>17</project.build.jdkVersion>
         <argLine>-Dfile.encoding=UTF-8</argLine>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <skip_maven_deploy>false</skip_maven_deploy>
@@ -66,12 +66,12 @@
            <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-checkstyle-plugin</artifactId>
-                <version>${maven-checkstyle-plugin-version}</version>
+                <version>3.2.0</version>
                 <dependencies>
                     <dependency>
                         <groupId>com.puppycrawl.tools</groupId>
                         <artifactId>checkstyle</artifactId>
-                        <version>8.18</version>
+                        <version>10.12.3</version>
                     </dependency>
                 </dependencies>
                 <executions>
@@ -80,7 +80,6 @@
                         <phase>validate</phase>
                         <configuration>
                             <configLocation>codestyle/checkstyle.xml</configLocation>
-                            <encoding>UTF-8</encoding>
                             <consoleOutput>true</consoleOutput>
                             <failOnViolation>true</failOnViolation>
                         </configuration>
