+++ dubbo-gateway-extensions/dubbo-gateway-provider/pom.xml
@@ -27,8 +27,8 @@
     <artifactId>dubbo-gateway-provider</artifactId>
 
     <properties>
-        <maven.compiler.source>8</maven.compiler.source>
-        <maven.compiler.target>8</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
     </properties>
