+++ dubbo-registry-extensions/dubbo-registry-nameservice/pom.xml
@@ -26,7 +26,20 @@
         <relativePath>../pom.xml</relativePath>
     </parent>
     <version>${revision}</version>
-	<modelVersion>4.0.0</modelVersion>
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <version>3.8.1</version>
+                <configuration>
+                    <source>17</source>
+                    <target>17</target>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
+    <modelVersion>4.0.0</modelVersion>
 	<artifactId>dubbo-registry-nameservice</artifactId>
 	<name>dubbo-registry-nameservice</name>
 	<properties>
