+++ dubbo-rpc-extensions/dubbo-rpc-native-thrift/pom.xml
@@ -52,5 +52,9 @@
             <artifactId>dubbo-remoting-zookeeper</artifactId>
             <optional>true</optional>
         </dependency>
+        <dependency>
+            <groupId>org.apache.dubbo</groupId>
+            <artifactId>dubbo</artifactId>
+        </dependency>
     </dependencies>
 </project>
