+++ dubbo-rpc-extensions/dubbo-rpc-rocketmq/pom.xml
@@ -26,6 +26,19 @@
         <relativePath>../pom.xml</relativePath>
     </parent>
     <version>${revision}</version>
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <version>3.8.1</version>
+                <configuration>
+                    <source>17</source>
+                    <target>17</target>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
     <artifactId>dubbo-rpc-rocketmq</artifactId>
     <name>dubbo-rpc-rocketmq</name>
     <properties>
