+++ dubbo-service-discovery-extensions/dubbo-service-discovery-springcloud/pom.xml
@@ -26,10 +26,11 @@
 
     <artifactId>dubbo-service-discovery-springcloud</artifactId>
     <version>${revision}</version>
+    <packaging>jar</packaging>
 
     <properties>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
