+++ dubbo-xds/src/main/java/org/apache/dubbo/registry/xds/util/protocol/impl/EdsProtocol.java
@@ -94,4 +94,9 @@ public class EdsProtocol extends AbstractProtocol<EndpointResult, DeltaEndpoint>
             return null;
         }
     }
+
+    @Override
+    public void updateObserve(long requestId, Set<String> resourceNames) {
+
+    }
 }
