+++ dubbo-xds/src/main/java/org/apache/dubbo/registry/xds/util/protocol/impl/LdsProtocol.java
@@ -103,4 +103,9 @@ public class LdsProtocol extends AbstractProtocol<ListenerResult, DeltaListener>
             return null;
         }
     }
+
+    @Override
+    public void updateObserve(long requestId, Set<String> resourceNames) {
+
+    }
 }
